+++ packages/gamut/src/ToolTip/PopoverToolTip.tsx
@@ -2,16 +2,17 @@ import React, { useLayoutEffect, useRef, useState } from 'react';
 
 import { Box, FlexBox } from '../Box';
 import { Popover } from '../Popover';
-import { ToolTipProps } from './types';
+import { tooltipDefaultProps, ToolTipProps } from './types';
 import { getPopoverAlignment } from './utils';
 
 export const PopoverToolTip: React.FC<ToolTipProps> = ({
-  alignment,
+  alignment = tooltipDefaultProps.alignment,
   children,
   focusable,
   id,
+  mode = tooltipDefaultProps.mode,
   target,
-  widthMode,
+  widthMode = tooltipDefaultProps.widthMode,
 }) => {
   const ref = useRef<HTMLDivElement>(null);
   const [offset, setOffset] = useState(0);
@@ -38,14 +39,15 @@ export const PopoverToolTip: React.FC<ToolTipProps> = ({
             (event.target as HTMLElement).blur();
           }
         }}
-        // ToolTips sometimes contain actual <button>s, which cannot be a child of a button.
-        // This element still needs tab focus so we must use the `tabIndex=0` hack.
-        // eslint-disable-next-line jsx-a11y/no-noninteractive-tabindex
         height="min-content"
         onBlur={() => setIsOpen(false)}
         onFocus={() => setIsOpen(true)}
         onMouseEnter={() => setIsOpen(true)}
         ref={ref}
+        role={focusable ? 'button' : undefined}
+        // ToolTips sometimes contain actual <button>s, which cannot be a child of a button.
+        // This element still needs tab focus so we must use the `tabIndex=0` hack.
+        // eslint-disable-next-line jsx-a11y/no-noninteractive-tabindex
         tabIndex={focusable ? 0 : undefined}
         width="min-content"
       >
@@ -62,6 +64,9 @@ export const PopoverToolTip: React.FC<ToolTipProps> = ({
         size="sml"
         targetRef={ref}
         widthRestricted={widthMode === 'standard'}
+        // TO-DO: mode
+        // bg={mode === 'light' ? 'white' : 'black'}
+        // color={mode === 'light' ? 'text' : 'secondary'}
       >
         <FlexBox alignItems="flex-start" flexDirection="column">
           {children}
